# library-management/src/main/java/com/library/library_management/auth/AuthenticationResponse.java
package com.library.library_management.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}